# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KESHAV",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/coverage": {
      "get": {
        "tags": [
          "coverage"
        ],
        "summary": "Get Coverage",
        "description": "Coverage per location across domains (weather, AQ, health, vulnerability).\n\nData-poor locations are reported as higher missingness \u2192 higher uncertainty.\nNo data is never treated as low risk.",
        "operationId": "get_coverage_api_v1_coverage_get",
        "parameters": [
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "name": "window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 720,
              "minimum": 1,
              "default": 48,
              "title": "Window Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoverageByLocation"
                  },
                  "title": "Response Get Coverage Api V1 Coverage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/explainability/prediction/{prediction_id}": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Prediction Explanation",
        "description": "Get local explanation for a specific prediction.\n\nReturns:\n    - Risk probability and baseline comparison\n    - Top positive and negative contributing factors\n    - Feature values and contributions\n    - In-distribution and OOD status\n    - Explanation quality validation",
        "operationId": "get_prediction_explanation_api_v1_explainability_prediction__prediction_id__get",
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prediction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Prediction Explanation Api V1 Explainability Prediction  Prediction Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explainability/global/{model_id}": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Global Explanation",
        "description": "Get global feature importance for a model.\n\nReturns:\n    - Ranked feature importance using specified method\n    - Total number of features\n    - Top contributing features",
        "operationId": "get_global_explanation_api_v1_explainability_global__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "shap",
                "permutation",
                "model_native"
              ],
              "type": "string",
              "default": "model_native",
              "title": "Method"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Global Explanation Api V1 Explainability Global  Model Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explainability/features": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Feature Manifest",
        "description": "Get the feature manifest for the currently deployed model.\n\nReturns:\n    - List of feature definitions with names, sources, units, and derivations",
        "operationId": "get_feature_manifest_api_v1_explainability_features_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Feature version filter",
              "title": "Version"
            },
            "description": "Feature version filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Feature Manifest Api V1 Explainability Features Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explainability/uncertainty/{prediction_id}": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Uncertainty",
        "description": "Get uncertainty metadata for a prediction.\n\nSeparates uncertainty from risk probability.\n\nReturns:\n    - Data quality assessment\n    - Data completeness score\n    - Missing/stale feature counts\n    - Forecast uncertainty (if applicable)\n    - Model uncertainty estimate\n    - Confidence category",
        "operationId": "get_uncertainty_api_v1_explainability_uncertainty__prediction_id__get",
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prediction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Uncertainty Api V1 Explainability Uncertainty  Prediction Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explainability/ood/{prediction_id}": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Ood Status",
        "description": "Get out-of-distribution status for a prediction.\n\nIdentifies features that are outside the model's training distribution.\n\nReturns:\n    - Whether prediction is in-distribution\n    - OOD score if applicable\n    - Features outside training range with distances\n    - Training range information\n    - OOD warning message",
        "operationId": "get_ood_status_api_v1_explainability_ood__prediction_id__get",
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prediction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Ood Status Api V1 Explainability Ood  Prediction Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explainability/equity/model/{model_id}": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Equity Audit",
        "description": "Get equity audit results for a model.\n\nEvaluates whether the model performs differently across subgroups.\n\nReturns:\n    - Subgroup performance metrics (recall, precision, FPR, FNR)\n    - Sample sizes and prevalence per subgroup\n    - Data coverage per subgroup\n    - Sufficiency flags for each subgroup",
        "operationId": "get_equity_audit_api_v1_explainability_equity_model__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Equity Audit Api V1 Explainability Equity Model  Model Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explainability/coverage": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Data Coverage",
        "description": "Get data coverage equity metrics across locations/wards.\n\nMeasures whether different areas have different data quality.\n\nReturns:\n    - Data coverage percentages per group/location\n    - Sensor availability, weather completeness, AQ completeness\n    - Forecast availability per area",
        "operationId": "get_data_coverage_api_v1_explainability_coverage_get",
        "parameters": [
          {
            "name": "subgroup",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Subgroup filter",
              "title": "Subgroup"
            },
            "description": "Subgroup filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Data Coverage Api V1 Explainability Coverage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explainability/model-audit/{model_id}": {
      "get": {
        "tags": [
          "explainability"
        ],
        "summary": "Get Model Audit",
        "description": "Get model audit records for a specific model.\n\nIncludes calibration diagnostics, explanation provenance,\nand overall model audit summary.\n\nReturns:\n    - Calibration diagnostics (Brier, ECE, reliability)\n    - Model metadata and version history\n    - Explanation provenance records",
        "operationId": "get_model_audit_api_v1_explainability_model_audit__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Model Audit Api V1 Explainability Model Audit  Model Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoverageByLocation": {
        "properties": {
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "domains": {
            "items": {
              "$ref": "#/components/schemas/CoverageReport"
            },
            "type": "array",
            "title": "Domains"
          }
        },
        "type": "object",
        "required": [
          "location_id"
        ],
        "title": "CoverageByLocation",
        "description": "Coverage rollup across locations \u2014 the basis for equity auditing."
      },
      "CoverageClass": {
        "type": "string",
        "enum": [
          "DATA_RICH",
          "ADEQUATE",
          "DATA_POOR"
        ],
        "title": "CoverageClass",
        "description": "Equity-oriented coverage labels (foundation for coverage auditing)."
      },
      "CoverageReport": {
        "properties": {
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "variable_coverage": {
            "items": {
              "$ref": "#/components/schemas/VariableCoverage"
            },
            "type": "array",
            "title": "Variable Coverage"
          },
          "spatial_resolution": {
            "$ref": "#/components/schemas/SpatialReference"
          },
          "temporal_resolution": {
            "$ref": "#/components/schemas/TemporalReference"
          },
          "overall_missingness_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Missingness Rate",
            "default": 0.0
          },
          "coverage_class": {
            "$ref": "#/components/schemas/CoverageClass",
            "default": "DATA_POOR"
          },
          "equity_note": {
            "type": "string",
            "title": "Equity Note",
            "default": "Data-poor locations yield higher uncertainty and lower confidence. No data is never treated as low risk."
          }
        },
        "type": "object",
        "required": [
          "location_id",
          "domain",
          "provider"
        ],
        "title": "CoverageReport",
        "description": "Spatial/temporal/variable coverage for one location.\n\nEncoding the principle: no data must never be interpreted as \"low risk\".\nMissing data raises uncertainty; low coverage yields lower confidence."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TemporalReference": {
        "properties": {
          "ref_type": {
            "$ref": "#/components/schemas/TimeWindowType",
            "default": "point"
          },
          "resolution_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Value"
          },
          "resolution_unit": {
            "$ref": "#/components/schemas/TemporalResolutionUnit",
            "default": "hour"
          }
        },
        "type": "object",
        "title": "TemporalReference",
        "description": "How a record is located in time (native resolution preserved)."
      },
...
      "TimeWindow": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "TimeWindow",
        "description": "An open/closed interval for alignment between different sources."
      },
      "TimeWindowType": {
        "type": "string",
        "enum": [
          "point",
          "window"
        ],
        "title": "TimeWindowType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}